Minor naming cleanup in reader
No functional changes
